Change forEach to .map
